add real time notification functionality
