+++ app/models/concerns/notifiable.rb
@@ -1,8 +1,12 @@
+require "action_view"
+require "action_view/helpers"
+
 module Notifiable
   extend ActiveSupport::Concern
 
   included do
-    include CableBroadcast
+    include ActionView::Helpers::DateHelper
+    include Acorn::NotificationBroadcast
     after_create :build_notification
   end
 
@@ -21,12 +25,34 @@ module Notifiable
         is_read: false
       )
 
-      # stream_content = {
-      #   id: id,
-      #   recipient: recipient[:username],
-      #   action: "#{user[:username]} commented on #{idea[:name]}",
-      # }
-      # send_broadcast(stream_content)
+      broadcast_content = {
+        actor: user.username,
+        topic: idea,
+        content: content,
+        picture: picture,
+        recipient: recipient[:username],
+        time: time,
+      }
+
+      send_broadcast(broadcast_content)
     end
+  end
+
+  private
+
+  def content
+    case self.class::ACTION.to_s
+    when "commented" then "commented on #{idea.name}"
+    else
+      "#{self.class::ACTION} #{idea.name}"
+    end
+  end
+
+  def picture
+    user.picture || "dummy-profile-male.jpg"
+  end
+
+  def time
+    "#{time_ago_in_words(created_at)} ago"
   end
 end
